login files are consolidated following the new file structure layout.
